VP-364: Moved export models from Data to Core project.
