+++ Modules/vc-module-catalog/VirtoCommerce.CatalogModule.Core/VirtoCommerce.CatalogModule.Core.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFramework>netcoreapp2.2</TargetFramework>
